Add content from LandingPage to CardSearchResult and left two buttons on LandingPage
